Sincronizar Devuelving CORE

+++ src/Models/ProductCustomModel.php
@@ -2,10 +2,17 @@
 
 namespace devuelving\core;
 
+use App\Product;
+use devuelving\core\FranchiseModel;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Builder;
+use Cviebrock\EloquentSluggable\Sluggable;
+use Cviebrock\EloquentSluggable\Services\SlugService;
 
 class ProductCustomModel extends Model
 {
+    use Sluggable;
+    
     /**
      * The table associated with the model.
      *
@@ -26,7 +33,7 @@ class ProductCustomModel extends Model
      * @var array
      */
     protected $fillable = [
-        'product', 'franchise', 'promotion', 'free_shipping', 'price', 'price_type', 'name', 'description', 'meta_title', 'meta_description', 'meta_keywords', 'removed',
+        'product', 'franchise', 'promotion', 'free_shipping', 'price', 'price_type', 'slug', 'name', 'description', 'meta_title', 'meta_description', 'meta_keywords', 'removed',
     ];
 
     /**
@@ -37,6 +44,27 @@ class ProductCustomModel extends Model
     protected $hidden = [
         'created_at', 'updated_at',
     ];
+
+    /**
+     * Return the sluggable configuration array for this model.
+     *
+     * @return array
+     */
+    public function sluggable()
+    {
+        return [
+            'slug' => [
+                'source' => ['slug'],
+                'unique' => true,
+                'onUpdate' => true,
+                'maxLength' => 191,
+            ]
+        ];
+    }
+    
+    public function scopeWithUniqueSlugConstraints(Builder $query, Model $model, $attribute, $config, $slug) {
+        return $query->where('franchise', FranchiseModel::getFranchise()->id);
+    }
 
     /**
      * The "booting" method of the model.
@@ -54,6 +82,37 @@ class ProductCustomModel extends Model
         self::updated(function($productCustom){
             $productCustom->checkClear();
         });
+
+        static::registerModelEvent('slugging', static function($model) {
+            //info('Product slugging: ' . $model->slug);
+
+            $product_slug = request('product_slug');
+            $slug_input = request('slug');
+
+            if(empty($model->slug) && empty($slug_input)) {
+                return false;
+            } else if(empty($slug_input) || !empty($slug_input) && $slug_input == $product_slug) {
+                return false;
+            }
+
+            $slug = SlugService::createSlug(ProductCustomModel::class, 'slug', $slug_input, ['unique' => false]);
+
+            // Existe SLUG de un producto con frnachise null
+            if(Product::where("id", "!=", $model->product)->whereNull("franchise")->where("slug", $slug)->exists()) {
+                return false;
+            }
+            // Existe SLUG de un producto customizado 
+            else if(ProductCustomModel::where("id", "!=", $model->id)->where("franchise", FranchiseModel::getFranchise()->id)->where("slug", $slug)->exists()) {
+                return false;
+            }
+
+            // Set slug
+            $model->slug = $slug;
+        });
+        
+        static::registerModelEvent('slugged', static function($model) {
+            //info('Category slugged: ' . $model->slug);
+        });                
     }
 
     /**
@@ -65,7 +124,7 @@ class ProductCustomModel extends Model
      */
     protected function checkClear()
     {
-        if ($this->promotion == null && $this->free_shipping == null && $this->price == null && $this->price_type == null && $this->name == null && $this->description == null && $this->meta_title == null && $this->meta_description == null && $this->meta_keywords == null && $this->removed == 0) {
+        if ($this->promotion == null && $this->free_shipping == null && $this->price == null && $this->price_type == null && $this->name == null && $this->description == null && $this->meta_title == null && $this->meta_description == null && $this->meta_keywords == null && $this->slug == null && $this->removed == 0 && $this->tags == null) {
             $this->delete();
         }
     }
